refactor: Update import paths for AlphabeticalSort in primary-options.js files
