+++ src/data/linux/primary-options.js
@@ -1,4 +1,4 @@
-import AlphabeticalSort from "utils/AlphabeticalSort";
+import AlphabeticalSort from "utils/alphabetical-sort";
 
 const options = [
   { value: 'list', label: 'list' },
